ingestion/requirements.txt
# Document processing
docling==2.54.0

# LlamaIndex and vector storage
llama-index==0.14.3
llama-index-vector-stores-postgres==0.7.0
llama-index-embeddings-ollama==0.8.3

# Database
psycopg2-binary==2.9.10
pgvector==0.4.1
sqlalchemy==2.0.43

# Ollama client
ollama==0.6.0

# CLI utilities
click==8.3.0
rich==14.1.0
python-dotenv==1.1.1

# Logging and configuration
pydantic==2.11.9
pydantic-settings==2.11.0

# Type hints
typing-extensions==4.15.0